Allow to post comment checklist at comment creation
