+++ zou/app/blueprints/tasks/__init__.py
@@ -12,7 +12,6 @@ from .resources import (
     PersonDoneTasksResource,
     TaskStartResource,
     ToReviewResource,
-    CommentTaskResource,
     TaskCommentsResource,
     TaskCommentResource,
     TaskPreviewsResource,
@@ -51,7 +50,6 @@ routes = [
     ("/data/projects/<project_id>/preview-files", ProjectPreviewFilesResource),
     ("/data/projects/<project_id>/subscriptions", ProjectSubscriptionsResource),
     ("/data/projects/<project_id>/tasks", ProjectTasksResource),
-    ("/actions/tasks/<task_id>/comment", CommentTaskResource),
     ("/actions/tasks/<task_id>/assign", TaskAssignResource),
     ("/actions/tasks/clear-assignation", ClearAssignationResource),
     ("/actions/persons/<person_id>/assign", TasksAssignResource),
